+++ app/controllers/application_controller.rb
@@ -90,4 +90,13 @@ class ApplicationController < ActionController::Base
   rescue
     nil
   end
+
+  # Expects an "Uploader" type class, ex: uploader = @project.logo
+  def send_file_if_present(uploader, *args)
+    if uploader.present?
+      send_file uploader.path, *args
+    else
+      head :ok
+    end
+  end
 end